[REFACTOR] Extract room export button
